fix embedding python test

+++ recipe/tests/prefix-replacement/build-and-test.sh
@@ -2,22 +2,30 @@
 
 set -x
 
-${CC} a.c $(python3-config --cflags) $(python3-config --embed --ldflags) -o ${CONDA_PREFIX}/bin/embedded-python-static
-if ${READELF} -d ${CONDA_PREFIX}/bin/embedded-python-static | rg libpython; then
-  echo "ERROR :: Embedded python linked to shared python library. It is expected to link to the static library."
+if [[ "$PKG_NAME" == "libpython-static" ]]; then
+  ${CC} a.c $(python3-config --cflags) $(python3-config --embed --ldflags) -o ${CONDA_PREFIX}/bin/embedded-python-static
+  if [[ -n ${READELF} ]]; then
+    if ${READELF} -d ${CONDA_PREFIX}/bin/embedded-python-static | rg libpython; then
+      echo "ERROR :: Embedded python linked to shared python library. It is expected to link to the static library."
+    fi
+  elif [[ -n ${OTOOL} ]]; then
+    if ${OTOOL} -l ${CONDA_PREFIX}/bin/embedded-python-static | rg libpython; then
+      echo "ERROR :: Embedded python linked to shared python library. It is expected to link to the static library."
+    fi
+  fi
+  ${CONDA_PREFIX}/bin/embedded-python-static
+
+  # I thought this would prefer the shared library for Python. I was wrong:
+  # EMBED_LDFLAGS=$(python3-config --ldflags)
+  # re='^(.*)(-lpython[^ ]*)(.*)$'
+  # if [[ ${EMBED_LDFLAGS} =~ $re ]]; then
+  #   EMBED_LDFLAGS="${BASH_REMATCH[1]} ${BASH_REMATCH[3]} -Wl,-Bdynamic ${BASH_REMATCH[2]}"
+  # fi
+  # ${CC} a.c $(python3-config --cflags) ${EMBED_LDFLAGS} -o ${CONDA_PREFIX}/bin/embedded-python-shared
+
+  # Brute-force way of linking to the shared library, sorry!
+  rm -rf ${CONDA_PREFIX}/lib/libpython*.a
 fi
-${CONDA_PREFIX}/bin/embedded-python-static
-
-# I thought this would prefer the shared library for Python. I was wrong:
-# EMBED_LDFLAGS=$(python3-config --ldflags)
-# re='^(.*)(-lpython[^ ]*)(.*)$'
-# if [[ ${EMBED_LDFLAGS} =~ $re ]]; then
-#   EMBED_LDFLAGS="${BASH_REMATCH[1]} ${BASH_REMATCH[3]} -Wl,-Bdynamic ${BASH_REMATCH[2]}"
-# fi
-# ${CC} a.c $(python3-config --cflags) ${EMBED_LDFLAGS} -o ${CONDA_PREFIX}/bin/embedded-python-shared
-
-# Brute-force way of linking to the shared library, sorry!
-rm -rf ${CONDA_PREFIX}/lib/libpython*.a
 
 ${CC} a.c $(python3-config --cflags) $(python3-config --embed --ldflags) -o ${CONDA_PREFIX}/bin/embedded-python-shared
 if [[ -n ${READELF} ]]; then
